MVKCmdDraw: Fix indirect index for triangle fan topology
When we call vkCmdDraw with vertexCount=N and firstVertex=X, we only
need an array with N element, X, X+1, ..., X+N-1, to generate the
indirect indices. However, the current logic generates all the way
from 0 to X+N-1 (which is X+N elements) while only reserving N spaces.
This causes issue when the unreserved part is overwritten, for example,
when issuing consecutive draw calls with triangle fan primitives.