MVKDevice: Support clustered subgroup ops at all sizes.
Recent changes in SPIRV-Cross allow us to support all cluster sizes, not
just 4. Update SPIRV-Cross to pull in this change.

Tested on Intel/AMD and Apple Silicon. The remaining failures, as usual,
are the 64-bit integer problem. I know how to solve this for the
permutation ops, but not the reduction ops.

Test results for `dEQP-VK.subgroups.clustered.*`:
Before:
  Passed:           0/3244 (0.0%)
  Failed:         268/3244 (8.26%)
  Not supported: 2976/3244 (91.74%)
  Warnings:         0/3244 (0.0%)
  Waived:           0/3244 (0.0%)
After:
  Passed:         212/3244 (6.54%)
  Failed:          56/3244 (1.73%)
  Not supported: 2976/3244 (91.74%)
  Warnings:         0/3244 (0.0%)
  Waived:           0/3244 (0.0%)